Update normalize_record after GH#3
This update copes with different structures for the "actual"
Update data when receiving Updates, with particular reference to
the CallbackQuery case as flagged by xeonvs in
https://github.com/polettix/Bot-ChatBots-Telegram/pull/3
"Fix for callback_query request from inline buttons".

The solution provided is more general than the pull request and
possibly more extensible in the future.